Added: mapService for Map-related Logic
