+++ Cluster/Cluster.py
@@ -6,6 +6,7 @@ from lux.game_map import Cell, Position
 from lux.constants import Constants
 from lux.game_constants import GAME_CONSTANTS
 from helperFunctions.helper_functions import *
+from Map.mapService import get_cell_neighbours_four, get_nearest_position
 
 from Resources.resourceService import get_resources_from_cells
