+++ Cluster/clusterController.py
@@ -7,8 +7,8 @@ from Cluster.Cluster import Cluster
 # from agent import game_state
 from lux.game_map import Cell
 
-from helperFunctions.helper_functions import (inside_map,
-    get_cell_neighbours_eight, same_resource)
+from helperFunctions.helper_functions import same_resource
+from Map.mapService import get_cell_neighbours_eight
 
 # logging.basicConfig(filename="ClusterController.log", level=logging.INFO)
